refactor(hooks) preserve route refactor during rebase
